LPS-195313 refactor upload handler classes on fragment-web
moving class to UploadFragmentCompositionPreviewMVCActionCommand.java, delete StyleBookEntryPreviewImageEditorUploadFileEntryHandler.java and change usages
